move index term UI into a separate view
