+++ scss/_sidebar.scss
@@ -210,6 +210,7 @@
                 grid-column: 1;
                 overflow: hidden;
                 font-size: 0.8em;
+                line-height: 2em;
             }
             .index-sdg {
                 padding: 0;
